% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excretion.R
\name{allocate_TAN_excretion}
\alias{allocate_TAN_excretion}
\title{allocate_TAN_excretion}
\usage{
allocate_TAN_excretion(exc_allocation, f_TAN)
}
\arguments{
\item{exc_allocation}{Nutrient excreted for a given pathway (yards, housing, grazing) (kg N/yr)}

\item{f_TAN}{(0-1) TAN fraction of manure}
}
\description{
Allocates manure excreted (in TAN) to the different systems
}
