% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Application.R
\name{application_NH3}
\alias{application_NH3}
\title{application_NH3}
\usage{
application_NH3(manure_appl_TAN, EF)
}
\arguments{
\item{manure_appl_TAN}{solid manure or slurry TAN application (kg TAN/yr)}

\item{EF}{NH3 EF for manure application (kg N-NH3/kg TAN)}
}
\description{
N-NH3 volatilisation during field application of solid manure or slurry
}
