% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{dairy_excretion_coeff}
\alias{dairy_excretion_coeff}
\title{dairy_excretion_coeff}
\usage{
dairy_excretion_coeff(milk_prod)
}
\arguments{
\item{milk_prod}{milk produced per dairy cow kg milk/(head.yr)}
}
\description{
updates dairy cattle nitrogen excreted (105) based on milk produced
}
\examples{
dairy_excretion_coeff(milk_prod = 10000)
}
