% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Digestates.R
\name{digested_application_N}
\alias{digested_application_N}
\title{digested_application_N}
\usage{
digested_application_N(biogas_slurry_N, biogas_solid_N, EF)
}
\arguments{
\item{biogas_slurry_N}{N flow from slurry for biogas (kg N/yr)}

\item{biogas_solid_N}{N flow from solid manure for biogas (kg N/yr}

\item{EF}{EF for N-NH3 emissions in storage}
}
\description{
calculates N in digestate that will be applied to the field
}
