% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Digestates.R
\name{digested_application_TAN}
\alias{digested_application_TAN}
\title{digested_application}
\usage{
digested_application_TAN(
  biogas_slurry_TAN,
  biogas_solid_TAN,
  biogas_slurry_N,
  biogas_solid_N,
  f_min_digester = 0.0067,
  EF
)
}
\arguments{
\item{biogas_slurry_TAN}{TAN flow from slurry for biogas (kg TAN/yr)}

\item{biogas_solid_TAN}{TAN flow from solid manure for biogas (kg TAN/yr)}

\item{biogas_slurry_N}{N flow from slurry for biogas (kg N/yr)}

\item{biogas_solid_N}{N flow from solid manure for biogas (kg N/yr}

\item{f_min_digester}{share of organic N entering the digester that is mineralized to TAN in the digester (0-1)}

\item{EF}{EF for N-NH3 emissions in storage}
}
\description{
calculates TAN in digestate that will be applied to the field
}
