% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bedding.R
\name{ex_housing_solid_N}
\alias{ex_housing_solid_N}
\title{ex_housing_solid_N}
\usage{
ex_housing_solid_N(housing_solid_N, housing_solid_NH3, animal_no, straw_N)
}
\arguments{
\item{housing_solid_N}{N deposited in housing from solid manure (kg N/head(head/yr))}

\item{housing_solid_NH3}{N-NH3 emissions in housing from solid manure (kg N-NH3/yr)}

\item{animal_no}{livestock number (head/yr)}

\item{straw_N}{N added in straw (kg N/(head.yr))}
}
\description{
Calculates the total amount of N in solid manure removed due to N-NH3 emissions
}
