% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bedding.R
\name{ex_housing_solid_TAN}
\alias{ex_housing_solid_TAN}
\title{ex_housing_solid_TAN}
\usage{
ex_housing_solid_TAN(
  housing_solid_TAN,
  housing_solid_NH3,
  animal_no,
  f_man_solid,
  bedding_amount,
  f_imm
)
}
\arguments{
\item{housing_solid_TAN}{TAN deposited in housing from solid manure (kg TAN/head(head/yr))}

\item{housing_solid_NH3}{N-NH3 emissions in housing from solid manure (kg N-NH3/yr)}

\item{animal_no}{livestock number (head/yr)}

\item{f_man_solid}{Proportion of manure handled as solid (0-1)}

\item{bedding_amount}{Annual straw use in bedding (kg N/(head.yr))}

\item{f_imm}{fraction of TAN immobilised in organic matter (0-1)}
}
\value{
Total TAN in solid manure including bedding TAN minus NH3 emissions
}
\description{
Calculates the total amount of TAN in solid manure removed from livestock due to immobilisation of TAN in bedding and NH3 emissions
}
\examples{
# Example usage
ex_housing_solid_TAN(100, 10, 50, 0.3, 500, 0.0067)
}
