% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grazing.R
\name{grazing_NH3}
\alias{grazing_NH3}
\title{grazing_NH3}
\usage{
grazing_NH3(grazing_tan, EF)
}
\arguments{
\item{grazing_tan}{total amount of TAN deposited onto pastures (kg TAN/yr)}

\item{EF}{Grazing emission factor (kg N-NH3/kg TAN)}
}
\description{
calculates N-NH3 losses from livestock yards
}
