% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Housing.R
\name{housing_deposited}
\alias{housing_deposited}
\title{housing_deposited}
\usage{
housing_deposited(housing_nutrient, f_man_type)
}
\arguments{
\item{housing_nutrient}{total amount of either N/TAN deposited onto animal housing (kg N/yr or kg TAN/yr)}

\item{f_man_type}{proportion of livestock manure handled as slurry or solid (0-1)}
}
\description{
amounts of N or TAN deposited during housing handled as either slurry or solid
}
