% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Housing.R
\name{housing_NH3}
\alias{housing_NH3}
\title{housing_NH3}
\usage{
housing_NH3(housing_tan, EF)
}
\arguments{
\item{housing_tan}{total amount of TAN deposited onto animal housing (kg TAN/yr)}

\item{EF}{Housing emission factor (kg N-NH3/kg TAN)}
}
\description{
calculates N-NH3 losses from livestock housing for both slurry and solid
}
