% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excretion.R
\name{livestock_excretion}
\alias{livestock_excretion}
\title{livestock_excretion}
\usage{
livestock_excretion(animal_no, exc_coef)
}
\arguments{
\item{animal_no}{livestock number (head /yr)}

\item{exc_coef}{nutrient excretion coefficient (kg nutrient/yr)}
}
\value{
int
}
\description{
calculates total nutrient excreted from livestock
}
\examples{
livestock_excretion(animal_no=50, exc_coef = 100)
}
