% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Application.R
\name{manure_direct_application}
\alias{manure_direct_application}
\title{manure_direct_application}
\usage{
manure_direct_application(out_storage, c_f_man_usage)
}
\arguments{
\item{out_storage}{output from storage_slurry (not updated_storage_slurry!) and storage_solid}

\item{c_f_man_usage}{}
}
\description{
manure_direct_application
}
