% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{mbc_1}
\alias{mbc_1}
\title{mbc_1
Mass balance check}
\usage{
mbc_1(exc_total, list_exc_allocation)
}
\arguments{
\item{exc_total}{total N excreted}

\item{list_exc_allocation}{list with N excreted onto different stages (housing, yards, grazing)}
}
\description{
mass balance check between total N excreted and N excreted onto different stages
}
\examples{
examples(10, c(1,3,6))
}
