% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grazing.R
\name{net_grazing}
\alias{net_grazing}
\title{net_manure_application}
\usage{
net_grazing(manure_grazing, grazing_NH3)
}
\arguments{
\item{manure_grazing}{Grazing N/TAN droppings (kg TAN/yr or kg N/yr)}

\item{grazing_NH3}{N-NH3 emissions from grazing (kg N-NH3/yr)}
}
\description{
computes net N/TAN amounts of manure excreted onto pastures
}
