% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Application.R
\name{net_manure_application}
\alias{net_manure_application}
\title{net_manure_application}
\usage{
net_manure_application(manure_appl, appl_NH3)
}
\arguments{
\item{manure_appl}{solid manure or slurry N/TAN application (kg TAN/yr or kg N/yr)}

\item{appl_NH3}{N-NH3 emissions from solid manure or slurry application (kg N-NH3/yr)}
}
\description{
computes net N/TAN amounts of solid manure or solid returned to the soil
}
