% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Application.R
\name{slurry_application}
\alias{slurry_application}
\title{slurry_application}
\usage{
slurry_application(
  appl_direct_slurry,
  out_storage_slurry,
  digestate,
  storage_em_slurry
)
}
\arguments{
\item{appl_direct_slurry}{TAN/N slurry with immediate application (kg TAN/yr or kg N/yr)}

\item{out_storage_slurry}{output from storage_slurry (not updated_storage_slurry!) and storage_solid (kg TAN/yr or kg N/yr)}

\item{digestate}{N or TAN with field application (kg TAN/yr or kg N/yr)}

\item{storage_em_slurry}{Total gaseous N emissions from slurry storage}
}
\description{
COmputes total N/TAN of slurry applied to the field
}
