% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Application.R
\name{solid_manure_application}
\alias{solid_manure_application}
\title{solid_manure_application}
\usage{
solid_manure_application(
  appl_direct_solid,
  out_storage_solid,
  storage_em_solid
)
}
\arguments{
\item{appl_direct_solid}{TAN/N solid manure with immediate application (kg TAN/yr or kg N/yr)}

\item{out_storage_solid}{output from storage_slurry (not updated_storage_slurry!) and storage_solid (kg TAN/yr or kg N/yr)}

\item{storage_em_solid}{Total gaseous N emissions from solid manure storage}
}
\description{
COmputes total N/TAN of solid manure applied to the field
}
