% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Storage.R
\name{storage_emissions}
\alias{storage_emissions}
\title{storage_emissions}
\usage{
storage_emissions(storage_TAN, EF_NH3, EF_N2O, EF_NO, EF_N2)
}
\arguments{
\item{storage_TAN}{TAN flows from slurry or solid manure in storage (kg TAN/yr)}

\item{EF_NH3}{Storage emission factor (kg N-NH3/kg TAN)}

\item{EF_N2O}{Storage emission factor (kg N-N2O/kg TAN)}

\item{EF_NO}{Storage emission factor (kg N-NO/kg TAN)}

\item{EF_N2}{Storage emission factor (kg N-N2/kg TAN)}
}
\value{
list for each one of the gases, and their total
}
\description{
computes storage gaseous losses (NH3, N2O, NO, N2) as well as their total
}
