% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Storage.R
\name{storage_slurry}
\alias{storage_slurry}
\title{storage_slurry}
\usage{
storage_slurry(
  housing_slurry,
  housing_slurry_NH3,
  yard,
  yard_NH3,
  f_man_usage_slurry
)
}
\arguments{
\item{housing_slurry}{TAN/N deposited in housing from slurry (kg TAN/yr or kg N/yr)}

\item{housing_slurry_NH3}{N-NH3 emissions in housing from slurry (kg N-NH3/yr)}

\item{yard}{TAN/N deposited in yards (kg TAN/yr or kg N/yr)}

\item{yard_NH3}{N-NH3 emissions in yards (kg N-NH3/yr)}

\item{f_man_usage_slurry}{Proportion of manure going to either biogas as feedstock OR stored must be (0-1)}
}
\description{
Estimates the amount of TAN/N from slurry either going to biogas as feedstock or ready for manure storage
}
