% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Storage.R
\name{storage_solid}
\alias{storage_solid}
\title{storage_solid}
\usage{
storage_solid(ex_housing_solid, f_man_usage_solid)
}
\arguments{
\item{ex_housing_solid}{N/TAN solid manure that leaves housing and passed to storage (kg TAN/yr or kg N/yr)}

\item{f_man_usage_solid}{Proportion of manure going to either biogas as feedstock OR stored must be (0-1)}
}
\description{
Estimates the amount of TAN/N from solid manure either going to biogas as feedstock or ready for manure storage
}
