% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Storage.R
\name{updated_storage_slurry}
\alias{updated_storage_slurry}
\title{updated_storage_slurry}
\usage{
updated_storage_slurry(out_storage_slurry_TAN, out_storage_slurry_N, f_min)
}
\arguments{
\item{out_storage_slurry_TAN}{TAN from slurry stored (kg TAN/yr)}

\item{out_storage_slurry_N}{N from slurry stored (kg N/yr)}

\item{f_min}{Fraction of organic N that is mineralised (0,1)}
}
\description{
Updates storage TAN from slurry to account for organic mineralisation of TAN before emission losses
}
