% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yards.R
\name{yards_NH3}
\alias{yards_NH3}
\title{yards_NH3}
\usage{
yards_NH3(yards_tan, EF)
}
\arguments{
\item{yards_tan}{total amount of TAN deposited onto yards (kg TAN/yr)}

\item{EF}{Yards emission factor (kg N-NH3/kg TAN)}
}
\description{
calculates N-NH3 losses from livestock yards
}
