% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bedding.R
\name{ex_housing_solid_N}
\alias{ex_housing_solid_N}
\title{ex_housing_solid_N}
\usage{
ex_housing_solid_N(
  housing_solid_N,
  housing_solid_NH3,
  animal_no,
  bedding_amount,
  f_man_solid
)
}
\arguments{
\item{housing_solid_N}{N deposited in housing from solid manure (kg N/head(head/yr))}

\item{housing_solid_NH3}{N-NH3 emissions in housing from solid manure (kg N-NH3/yr)}

\item{animal_no}{livestock number (head/yr)}

\item{bedding_amount}{Annual straw use in bedding (kg N/(head.yr))}

\item{f_man_solid}{Proportion of manure handled as solid (0-1)}
}
\description{
Calculates the total amount of N in solid manure removed due to N-NH3 emissions
}
